# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Homeworking API",
    "description": "Bauvorhaben als veränderbares, berechenbares Projektmodell.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "operationId": "list_projects",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectListEntry"
                  },
                  "type": "array",
                  "title": "Response List Projects"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "operationId": "create_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Project",
        "operationId": "delete_project",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/commands": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Execute Command",
        "operationId": "execute_command",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/undo": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Undo",
        "operationId": "undo",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/history": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "History",
        "operationId": "history",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HistoryEntry"
                  },
                  "title": "Response History"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/versions": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Versions",
        "description": "Every log entry is a version that can be viewed and restored (newest first).",
        "operationId": "versions",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VersionEntry"
                  },
                  "title": "Response Versions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/versions/{seq}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Version",
        "description": "Read-only view of the project as it was after version ``seq``.",
        "operationId": "get_version",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "seq",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Seq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/versions/{seq}/restore": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Restore Version",
        "operationId": "restore_version",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "seq",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Seq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/drawings/{view}.svg": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Drawing",
        "operationId": "drawing",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "view",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "View"
            }
          },
          {
            "name": "seq",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "image/svg+xml": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/document.pdf": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Document",
        "operationId": "document",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "seq",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/pdf": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/export": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Export",
        "operationId": "export",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Export"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommandRequest": {
        "properties": {
          "command": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/SetParameters"
              },
              {
                "$ref": "#/components/schemas/ChangeParameterBy"
              },
              {
                "$ref": "#/components/schemas/SelectVariant"
              },
              {
                "$ref": "#/components/schemas/Rename"
              },
              {
                "$ref": "#/components/schemas/SetRegion"
              },
              {
                "$ref": "#/components/schemas/AddNote"
              },
              {
                "$ref": "#/components/schemas/ReplaceDesign"
              },
              {
                "$ref": "#/components/schemas/SetPrice"
              }
            ],
            "title": "Command",
            "discriminator": {
              "propertyName": "type",
              "mapping": {
                "add_note": "#/components/schemas/AddNote",
                "change_parameter_by": "#/components/schemas/ChangeParameterBy",
                "rename": "#/components/schemas/Rename",
                "replace_design": "#/components/schemas/ReplaceDesign",
                "select_variant": "#/components/schemas/SelectVariant",
                "set_parameters": "#/components/schemas/SetParameters",
                "set_price": "#/components/schemas/SetPrice",
                "set_region": "#/components/schemas/SetRegion"
              }
            }
          }
        },
        "type": "object",
        "required": [
          "command"
        ],
        "title": "CommandRequest"
      },
      "CommandResponse": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectModel"
          },
          "can_undo": {
            "type": "boolean",
            "title": "Can Undo"
          },
          "diff": {
            "$ref": "#/components/schemas/ModelDiff"
          },
          "seq": {
            "type": "integer",
            "title": "Seq"
          }
        },
        "type": "object",
        "required": [
          "project",
          "can_undo",
          "diff",
          "seq"
        ],
        "title": "CommandResponse"
      },
...
      "CreateProjectRequest": {
        "properties": {
          "pack_id": {
            "type": "string",
            "title": "Pack Id",
            "description": "Construction pack id, 'design' for a free-form design or 'template'"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "params": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "string"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "type": "object",
            "title": "Params"
          },
          "region": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Region"
              },
              {
                "type": "null"
              }
            ]
          },
          "design": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssemblyDesign"
              },
              {
                "type": "null"
              }
            ],
            "description": "Required for pack_id 'design'"
          },
          "template_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Key",
            "description": "Required for pack_id 'template'"
          }
        },
        "type": "object",
        "required": [
          "pack_id",
          "title"
        ],
        "title": "CreateProjectRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistoryEntry": {
        "properties": {
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "actor": {
            "type": "string",
            "title": "Actor"
          },
          "command": {
            "additionalProperties": true,
            "type": "object",
            "title": "Command"
          },
          "engine_version": {
            "type": "string",
            "title": "Engine Version"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "seq",
          "actor",
          "command",
          "engine_version",
          "created_at"
        ],
        "title": "HistoryEntry"
      },
...
      "ProjectListEntry": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "pack_id": {
            "type": "string",
            "title": "Pack Id"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "pack_id",
          "summary",
          "updated_at"
        ],
        "title": "ProjectListEntry"
      },
      "ProjectModel": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "inputs": {
            "$ref": "#/components/schemas/ProjectInputs"
          },
          "result": {
            "$ref": "#/components/schemas/ConstructionResult"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "inputs",
          "result"
        ],
        "title": "ProjectModel",
        "description": "Single source of truth of a construction project."
      },
      "ProjectView": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectModel"
          },
          "can_undo": {
            "type": "boolean",
            "title": "Can Undo"
          }
        },
        "type": "object",
        "required": [
          "project",
          "can_undo"
        ],
        "title": "ProjectView"
      },
...
      "Region": {
        "type": "string",
        "enum": [
          "BW",
          "BY",
          "BE",
          "BB",
          "HB",
          "HH",
          "HE",
          "MV",
          "NI",
          "NW",
          "RP",
          "SL",
          "SN",
          "ST",
          "SH",
          "TH"
        ],
        "title": "Region",
        "description": "German federal states; building regulations differ per state."
      },
...
    }
  }
}